+++ clang/lib/Parse/ParseDecl.cpp
@@ -5354,17 +5354,20 @@ void Parser::ParseEnumSpecifier(SourceLocation StartLoc, DeclSpec &DS,
 
   bool Owned = false;
   bool IsDependent = false;
+  TypeResult MSVCEnumType;
   const char *PrevSpec = nullptr;
   unsigned DiagID;
   Decl *TagDecl =
-      Actions.ActOnTag(getCurScope(), DeclSpec::TST_enum, TUK, StartLoc, SS,
-                    Name, NameLoc, attrs, AS, DS.getModulePrivateSpecLoc(),
-                    TParams, Owned, IsDependent, ScopedEnumKWLoc,
-                    IsScopedUsingClassTag,
+      Actions
+          .ActOnTag(getCurScope(), DeclSpec::TST_enum, TUK, StartLoc, SS, Name,
+                    NameLoc, attrs, AS, DS.getModulePrivateSpecLoc(), TParams,
+                    Owned, IsDependent, ScopedEnumKWLoc, IsScopedUsingClassTag,
                     BaseType, DSC == DeclSpecContext::DSC_type_specifier,
                     DSC == DeclSpecContext::DSC_template_param ||
                         DSC == DeclSpecContext::DSC_template_type_arg,
-                    OffsetOfState, &SkipBody).get();
+                    OffsetOfState, &SkipBody, &MSVCEnumType,
+                    DS.isFriendSpecified())
+          .get();
 
   if (SkipBody.ShouldSkip) {
     assert(TUK == TagUseKind::Definition && "can only skip a definition");
@@ -5403,6 +5406,15 @@ void Parser::ParseEnumSpecifier(SourceLocation StartLoc, DeclSpec &DS,
                            Actions.getASTContext().getPrintingPolicy()))
       Diag(StartLoc, DiagID) << PrevSpec;
 
+    return;
+  }
+
+  if (MSVCEnumType.isUsable()) {
+    if (DS.SetTypeSpecType(DeclSpec::TST_typename, StartLoc,
+                           NameLoc.isValid() ? NameLoc : StartLoc, PrevSpec,
+                           DiagID, MSVCEnumType.get(),
+                           Actions.getASTContext().getPrintingPolicy()))
+      Diag(StartLoc, DiagID) << PrevSpec;
     return;
   }